+++ customized_scripts/L_TS_ribo-bwamem.sh
@@ -19,42 +19,78 @@ out=$4
 
 source $general_parameter_filepath
 source $run_parameter_filepath
+
 current_dir=$(pwd)
-echo "Going to $outdir"
-cd $outdir
+
+if [[ $TMPDIR == "" ]]; then
+  echo "Going to $outdir"
+  cd $outdir
+else
+  echo "Going to $TMPDIR"
+  cd $TMPDIR
+fi
 
 # mapping short reads
 echo "mapping rRNA, short"
 ${p2bwa}/bwa aln ${riboref} ${fq} > aln_${fq%.f*q}.sai 
-${p2bwa}/bwa samse  ${riboref}  aln_${fq%.f*q}.sai ${fq} > aln-temp.out 
-${p2samtools}/samtools view -Sb aln-temp.out  > ${out}.aln-ribo.bam 
-rm aln-temp.out # i don't know why, but locally it only worked with creating an intermediate file
+${p2bwa}/bwa samse  ${riboref}  aln_${fq%.f*q}.sai ${fq} | ${p2samtools}/samtools view -Sb  > ${out}.aln-ribo.bam &
+  # Code for running stuff locally:
+  #${p2bwa}/bwa samse  ${riboref}  aln_${fq%.f*q}.sai ${fq} > aln-temp.out 
+  #${p2samtools}/samtools view -Sb aln-temp.out  > ${out}.aln-ribo.bam 
+  #rm aln-temp.out # i don't know why, but locally it only worked with creating an intermediate file
 
 # mapping normal reads
 echo "mapping rRNA, long"
-${p2bwa}/bwa mem -t 8 -h 15 ${riboref} ${fq} > mem-temp.out
-${p2samtools}/samtools view -Sb mem-temp.out > ${out}.mem-ribo.bam 
-rm mem-temp.out
+${p2bwa}/bwa mem -t 8 -h 15 ${riboref} ${fq} | ${p2samtools}/samtools view -Sb > ${out}.mem-ribo.bam 
+  # Code for running stuff locally:
+  #${p2bwa}/bwa mem -t 8 -h 15 ${riboref} ${fq} > mem-temp.out
+  #${p2samtools}/samtools view -Sb mem-temp.out > ${out}.mem-ribo.bam 
+  #rm mem-temp.out
+
+# (Wait since above processes are both run in the background..)
+wait
 
 echo "merging two maps" 
-${p2samtools}/samtools merge -f -n -r -h ${out}.aln-ribo.bam ${out}.all-ribo.bam ${out}.aln-ribo.bam ${out}.mem-ribo.bam 
-${p2samtools}/samtools view -H ${out}.aln-ribo.bam
+${p2samtools}/samtools merge -f -n -r -h ${out}.aln-ribo.bam --threads 8 ${out}.all-ribo.bam ${out}.aln-ribo.bam ${out}.mem-ribo.bam 
+  # differences with anna: -f option (overwrite output if already exists)
+  # note: -h indicates the file from which to take headers
+  # note: to debug, can be useful to convert to sam, with 
+  # out=SRR6640430_nc_cbc_trimmed_HATCG
+  # p2samtools=/hpc/hub_oudenaarden/mwehrens/bin/miniconda3/bin
+  # ${p2samtools}/samtools view -h -o ${out}.aln-ribo.sam ${out}.aln-ribo.bam
+# remove input files
+if [[ -f ${out}.all-ribo.bam ]]; then
+  echo "merge complete"
+  rm ${out}.aln-ribo.bam ${out}.mem-ribo.bam aln_${fq%.f*q}.sai
+else
+  echo "something failed during ribo merge"
+  exit 1
+fi
+  
+  
 
-rm ${out}.aln-ribo.bam ${out}.mem-ribo.bam aln_${fq%.f*q}.sai
+# Not sure why I added this .. (it prints headers)
+# ${p2samtools}/samtools view -H ${out}.aln-ribo.bam
 
 echo "sorting"
-${p2samtools}/samtools sort -n -O bam -T temp_${out} -o ${out}.nsorted.all-ribo.bam ${out}.all-ribo.bam
+${p2samtools}/samtools sort -n --threads 8 -O bam -T temp_${out} -o ${out}.nsorted.all-ribo.bam ${out}.all-ribo.bam
+  # note: -T indicates where to write temporary files
 rm ${out}.all-ribo.bam
 
 echo "riboread selection"
 #pythonbin=/Users/m.wehrens/anaconda3/bin/python
 $pythonbin ${p2s}/TS_riboread-selection.py ${out}.nsorted.all-ribo.bam $stranded ${out}
 
-if [[ $nocleanup = "" ]]; then
-  echo "cleaning up some files" # prevent this by setting "nocleanup"
-  rm ${fq}
-  rm ${out}.nsorted.all-ribo.bam
-  rm ${out}.Ribo.bam
+if [[ -f ${out}.nonRibo.fastq.gz ]]; then
+  if [[ $nocleanup = "" ]]; then
+    echo "cleaning up some files" # prevent this by setting "nocleanup"
+    rm ${fq}
+    rm ${out}.nsorted.all-ribo.bam
+    rm ${out}.Ribo.bam
+  fi
+else
+    echo "Expected output of this script not detected."
+    exit 1
 fi
 
 echo "end of script"
